refactor : Option directory 분리 및 변수 및 형식 통일
수정 :
- directory 분리하여, 객체들간의 의존을 확실하게 보이게 하였습니다.
- 변수 이름을 통일하였습니다.
- 형식을 통일하였습니다.
