+++ src/main/java/subway/controller/ShortestPathFinder.java
@@ -6,7 +6,6 @@ import subway.domain.StationRepository;
 import subway.view.PathInputView;
 
 abstract public class ShortestPathFinder {
-
     private final Scanner scanner;
 
     ShortestPathFinder(Scanner scanner) {
