+++ src/main/java/subway/controller/option/PathOption.java
@@ -1,6 +1,8 @@
-package subway.controller;
+package subway.controller.option;
 
 import java.util.Scanner;
+import subway.controller.ShortestDistancePathFinder;
+import subway.controller.ShortestTimePathFinder;
 import subway.controller.constants.QuestionNumber;
 
 public enum PathOption {
